Refactor: 금액 관련 타입 Long타입으로 수정

+++ module-api/src/main/java/com/example/moduleapi/controller/request/auction/BidRequest.java
@@ -6,9 +6,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class BidRequest {
-    private int biddingPrice;
+    private Long biddingPrice;
 
-    public BidRequest(int biddingPrice) {
+    public BidRequest(Long biddingPrice) {
         this.biddingPrice = biddingPrice;
     }
 }
