+++ module-api/src/test/java/com/example/moduleapi/fixture/PointFixtures.java
@@ -6,10 +6,10 @@ import com.example.moduleapi.controller.response.point.PointResponse;
 public class PointFixtures {
 
     public static PointAmount 포인트_충전 = PointAmount.builder()
-                                                  .amount(100000)
+                                                  .amount(100000L)
                                                   .build();
 
-    public static PointResponse 포인트_응답(int chargedAmount, int currentBalance) {
+    public static PointResponse 포인트_응답(Long chargedAmount, Long currentBalance) {
         String message = String.format("포인트 %d원 충전 완료. [현재 포인트 잔액 : %d원]", chargedAmount, currentBalance);
         return PointResponse.from(currentBalance, message);
     }
